+++ dokdo/cli/__init__.py
@@ -3,7 +3,6 @@ from .make_manifest import make_manifest
 from .add_metadata import add_metadata
 from .summarize import summarize
 from .prepare_lefse import prepare_lefse
-from .count_reads import count_reads
 
 commands = {
     "collapse": collapse,
@@ -11,5 +10,4 @@ commands = {
     "add-metadata": add_metadata,
     "summarize": summarize,
     "prepare-lefse": prepare_lefse,
-    "count-reads": count_reads,
 }