+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 -----------------------
 
 * :issue:`14`: Add the ``group_order`` option to the :meth:`taxa_abundance_bar_plot` method.
+* Fix a minor bug in the :meth:`addbiplot` method when feature is 'Unassigned'.
 
 1.9.0 (2021-06-07)
 ----------------------