+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 * :issue:`14`: Add the ``group_order`` option to the :meth:`taxa_abundance_bar_plot` method.
 * Fix a minor bug in the :meth:`addbiplot` method when feature is 'Unassigned'.
+* Deprecate the :command:`count-reads` command.
 
 1.9.0 (2021-06-07)
 ----------------------