+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 Changelog
 *********
 
-1.10.0 (in development)
------------------------
+1.10.0 (2021-07-06)
+-------------------
 
 * :issue:`14`, :issue:`17`: Add the ``group_order`` option to the :meth:`taxa_abundance_bar_plot` method.
 * Fix a minor bug in the :meth:`addbiplot` method when feature is 'Unassigned'.
@@ -13,7 +13,7 @@ Changelog
   - :meth:`read_quality_plot`
 
 1.9.0 (2021-06-07)
-----------------------
+------------------
 
 * Add publicly available datasets from QIIME 2 for tutorials.
 * :issue:`14`: Add the ``group`` option to the :meth:`taxa_abundance_bar_plot` method. Using this option will create a bar for each group instead of each sample.